requirements.txt
django~=5.1.4
anthropic~=0.40.0
python-dotenv~=1.0.1
langchain~=0.3.10
langchain-anthropic~=0.3.0
langchain-community~=0.3.10
langchain-openai~=0.2.12
openai~=1.57.1
PyPDF2~=3.0.1
chromadb
tiktoken
requests~=2.32.3
django-environ~=0.11.2
PyMuPDF
psycopg2-binary
whitenoise
gunicorn
python-decouple
django-storages[azure]
django-cors-headers
openparse[ml]~=0.7.0
google-search-results
aiohttp~=3.9.3
uvicorn[standard]~=0.27.0
azure-storage-blob==12.23.1
azure-core~=1.32.0
django-storages~=1.14.4
crewai~=0.11.0
crewai-tools~=0.0.10
serpapi~=0.1.5
